Moved DestinationTestError as inline component + updated tests
